Checkpoint: mobile app, auth + balances handlers, Swish design doc
This is a single WIP checkpoint commit so the in-progress Week 8.5 work and
the Swish integration design doc are on main before fanning out Phase 2
implementation in parallel worktrees.

Includes:
- Expo / React Native mobile app (sign-in, tabs, groups, expenses,
  onboarding, settle flow scaffolding) under app/
- Backend handlers: auth (magic-link + JWT), activity, balances, settle,
  settle-suggestions; settle algorithm in internal/settle/
- Migration 000012 — balance view with settlement offsets (CTE)
- sqlc regen for activity, group_members, models
- expenses handler / config / server tweaks
- docs/implementation-status.md tracking through Week 8.5
- docs/swish-integration.md — full end-to-end design for Swish
  (deep-link + callbackurl flow, optional merchant tier in §9,
  7-PR implementation plan in §8)
- docs/design-system/ — UI kits and previews
- backend/.env.dev with dev-only credentials, .air.toml,
  backend/.gitignore
- Root .gitignore for .claude/, .gstack/, docs/superpowers/

To be reorganized into proper per-feature commits later; for now this
unblocks parallel-worktree implementation of the Swish PRs.

+++ backend/internal/db/models.go
@@ -91,7 +91,7 @@ type MemberBalance struct {
 	MemberID   string      `db:"member_id" json:"member_id"`
 	UserID     pgtype.Text `db:"user_id" json:"user_id"`
 	Currency   pgtype.Text `db:"currency" json:"currency"`
-	NetBalance int32       `db:"net_balance" json:"net_balance"`
+	NetBalance int64       `db:"net_balance" json:"net_balance"`
 }
 
 type PushToken struct {
